Main class LobeGenerator for stat and non stat classes
